+++ lib/Extension/Core/CoreExtension.php
@@ -18,8 +18,8 @@ use Phpactor\ClassFileConverter\Domain\ChainFileToClass;
 use Phpactor\ClassFileConverter\Domain\ClassToFileFileToClass;
 use Phpactor\ClassMover\ClassMover;
 use Phpactor\Filesystem\Domain\Cwd;
-use Phpactor\Console\Command\ClassCopyCommand;
-use Phpactor\Console\Command\ClassMoveCommand;
+use Phpactor\Extension\ClassMover\Command\ClassCopyCommand;
+use Phpactor\Extension\ClassMover\Command\ClassMoveCommand;
 use Phpactor\Console\Command\ClassReflectorCommand;
 use Phpactor\Console\Command\ClassSearchCommand;
 use Phpactor\Console\Command\OffsetInfoCommand;
@@ -35,10 +35,6 @@ use Phpactor\Console\Command\ConfigDumpCommand;
 use Monolog\Logger;
 use Phpactor\Application\Complete;
 use Phpactor\Console\Command\CompleteCommand;
-use Phpactor\Extension\ClassMover\Application\ClassReferences;
-use Phpactor\Console\Command\ReferencesClassCommand;
-use Phpactor\Console\Command\ReferencesMemberCommand;
-use Phpactor\Extension\ClassMover\Application\ClassMemberReferences;
 use Psr\Log\LogLevel;
 use Monolog\Handler\StreamHandler;
 use Monolog\Handler\FingersCrossedHandler;
@@ -97,7 +93,6 @@ class CoreExtension implements Extension
         $this->registerConsole($container);
         $this->registerComposer($container);
         $this->registerClassToFile($container);
-        $this->registerClassMover($container);
         $this->registerApplicationServices($container);
     }
 
@@ -130,20 +125,6 @@ class CoreExtension implements Extension
         // ---------------
         // Commands
         // ---------------
-        $container->register('command.class_move', function (Container $container) {
-            return new ClassMoveCommand(
-                $container->get('application.class_mover'),
-                $container->get('console.prompter')
-            );
-        }, [ 'ui.console.command' => []]);
-
-        $container->register('command.class_copy', function (Container $container) {
-            return new ClassCopyCommand(
-                $container->get('application.class_copy'),
-                $container->get('console.prompter')
-            );
-        }, [ 'ui.console.command' => []]);
-
         $container->register('command.class_search', function (Container $container) {
             return new ClassSearchCommand(
                 $container->get('application.class_search'),
@@ -183,20 +164,6 @@ class CoreExtension implements Extension
         $container->register('command.complete', function (Container $container) {
             return new CompleteCommand(
                 $container->get('application.complete'),
-                $container->get('console.dumper_registry')
-            );
-        }, [ 'ui.console.command' => []]);
-
-        $container->register('command.class_references', function (Container $container) {
-            return new ReferencesClassCommand(
-                $container->get('application.class_references'),
-                $container->get('console.dumper_registry')
-            );
-        }, [ 'ui.console.command' => []]);
-
-        $container->register('command.method_references', function (Container $container) {
-            return new ReferencesMemberCommand(
-                $container->get('application.method_references'),
                 $container->get('console.dumper_registry')
             );
         }, [ 'ui.console.command' => []]);
@@ -333,50 +300,8 @@ class CoreExtension implements Extension
         });
     }
 
-    private function registerClassMover(ContainerBuilder $container)
-    {
-        $container->register('class_mover.class_mover', function (Container $container) {
-            return new ClassMover(
-                $container->get('class_mover.class_finder'),
-                $container->get('class_mover.ref_replacer')
-            );
-        });
-
-        $container->register('class_mover.class_finder', function (Container $container) {
-            return new \Phpactor\ClassMover\Adapter\TolerantParser\TolerantClassFinder();
-        });
-
-        $container->register('class_mover.member_finder', function (Container $container) {
-            return new \Phpactor\ClassMover\Adapter\WorseTolerant\WorseTolerantMemberFinder($container->get('reflection.reflector'));
-        });
-
-        $container->register('class_mover.member_replacer', function (Container $container) {
-            return new \Phpactor\ClassMover\Adapter\WorseTolerant\WorseTolerantMemberReplacer();
-        });
-
-        $container->register('class_mover.ref_replacer', function (Container $container) {
-            return new \Phpactor\ClassMover\Adapter\TolerantParser\TolerantClassReplacer();
-        });
-    }
-
     private function registerApplicationServices(Container $container)
     {
-        $container->register('application.class_mover', function (Container $container) {
-            return new ClassMoverApp(
-                $container->get('application.helper.class_file_normalizer'),
-                $container->get('class_mover.class_mover'),
-                $container->get('source_code_filesystem.registry')
-            );
-        });
-
-        $container->register('application.class_copy', function (Container $container) {
-            return new ClassCopy(
-                $container->get('application.helper.class_file_normalizer'),
-                $container->get('class_mover.class_mover'),
-                $container->get('source_code_filesystem.registry')->get('git')
-            );
-        });
-
         $container->register('application.file_info', function (Container $container) {
             return new FileInfo(
                 $container->get('class_to_file.converter'),
@@ -410,25 +335,6 @@ class CoreExtension implements Extension
             return new Complete(
                 $container->get('completion.completor'),
                 $container->get('application.helper.class_file_normalizer')
-            );
-        });
-
-        $container->register('application.class_references', function (Container $container) {
-            return new ClassReferences(
-                $container->get('application.helper.class_file_normalizer'),
-                $container->get('class_mover.class_finder'),
-                $container->get('class_mover.ref_replacer'),
-                $container->get('source_code_filesystem.registry')
-            );
-        });
-
-        $container->register('application.method_references', function (Container $container) {
-            return new ClassMemberReferences(
-                $container->get('application.helper.class_file_normalizer'),
-                $container->get('class_mover.member_finder'),
-                $container->get('class_mover.member_replacer'),
-                $container->get('source_code_filesystem.registry'),
-                $container->get('reflection.reflector')
             );
         });
